Silence activity stop keyboard shortcut
This patch is backported from master.

The Sugar keyboard shortcut for activity stop is control-q.

MusicKeyboard-8.1 does sound a spurious note briefly when the shortcut
is used.

Add an event state modifier mask to ignore key presses where the control
key is held down.

Also fix a traceback that occurs when a key release event arrives
without a corresponding key press event.  This can also happen if the
key press event is delivered to another activity or the shell.

Signed-off-by: James Cameron <quozl@laptop.org>